requirements.txt
# Core Data Processing
pandas>=1.5.0
numpy>=1.21.0

# Database Connectivity
psycopg2-binary>=2.9.0
SQLAlchemy>=1.4.0

# Data Generation and Simulation
Faker>=15.0.0
scipy>=1.9.0

# Configuration Management
PyYAML>=6.0
python-dotenv>=0.19.0

# Logging and Utilities
colorlog>=6.6.0
tqdm>=4.64.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Data Validation
Great-Expectations>=0.15.0

# API and Web (if needed for data ingestion)
requests>=2.28.0

# Date/Time Utilities
python-dateutil>=2.8.0

# Visualization (for data exploration during development)
matplotlib>=3.5.0
seaborn>=0.11.0

# Performance and Memory Management
memory-profiler>=0.60.0

# Code Quality
black>=22.0.0
flake8>=4.0.0

# Development Dependencies
jupyter>=1.0.0
ipython>=8.0.0 